backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
SQLAlchemy==2.0.38
asyncpg==0.30.0
alembic==1.14.1
python-multipart==0.0.20
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.1
httpx==0.28.1
